src/word2vec/requirements.txt
mysql-connector-python
pandas
gensim